Extract settings.ts
This moves all setting related functions to a new file called settings.ts. This is in preparation for a more formal settings infrastructure.
