+++ src/extension.ts
@@ -11,17 +11,18 @@ import {
   CloseAction,
   ErrorAction,
 } from 'vscode-languageclient/node';
-import { getInitializationOptions, getServerOptions } from './utils/clientHelpers';
+import { getServerOptions } from './utils/clientHelpers';
 import { GenerateBugReportCommand } from './commands/generateBugReport';
 import { ModuleCallsDataProvider } from './providers/moduleCalls';
 import { ModuleProvidersDataProvider } from './providers/moduleProviders';
 import { ServerPath } from './utils/serverPath';
-import { config, deleteSetting, getScope, migrate, warnIfMigrate } from './utils/vscode';
+import { config, getScope } from './utils/vscode';
 import { TelemetryFeature } from './features/telemetry';
 import { ShowReferencesFeature } from './features/showReferences';
 import { CustomSemanticTokens } from './features/semanticTokens';
 import { ModuleProvidersFeature } from './features/moduleProviders';
 import { ModuleCallsFeature } from './features/moduleCalls';
+import { getInitializationOptions, migrateLegacySettings, previewExtensionPresent } from './settings';
 
 const id = 'terraform';
 const brand = `HashiCorp Terraform`;
@@ -44,7 +45,6 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
     return undefined;
   }
 
-  // migrate pre-2.24.0 settings
   await migrateLegacySettings(context);
 
   // Subscriptions
@@ -256,117 +256,4 @@ async function stopLanguageServer() {
 
 function enabled(): boolean {
   return config('terraform').get('languageServer.enable', false);
-}
-
-async function migrateLegacySettings(ctx: vscode.ExtensionContext) {
-  // User has asked not to check if settings need to be migrated, so return
-  if (ctx.globalState.get('terraform.disableSettingsMigration', false)) {
-    return;
-  }
-
-  // If any of the following list needs to be migrated, ask user if they want
-  // to migrate. This is a blunt force approach, but we don't intend to keep
-  // checking this forever
-  const warnMigration = warnIfMigrate([
-    { section: 'terraform', name: 'languageServer.external' },
-    { section: 'terraform', name: 'languageServer.pathToBinary' },
-    { section: 'terraform-ls', name: 'rootModules' },
-    { section: 'terraform-ls', name: 'excludeRootModules' },
-    { section: 'terraform-ls', name: 'ignoreDirectoryNames' },
-    { section: 'terraform-ls', name: 'terraformExecPath' },
-    { section: 'terraform-ls', name: 'terraformExecTimeout' },
-    { section: 'terraform-ls', name: 'terraformLogFilePath' },
-    { section: 'terraform-ls', name: 'experimentalFeatures' },
-  ]);
-  if (warnMigration === false) {
-    return;
-  }
-
-  const messageText =
-    'Automatic migration will change your settings file!' +
-    '\n\nTo read more about the this change click "More Info" and delay changing anything';
-  // Prompt the user if they want to migrate. If the choose no, then return
-  // and they are left to migrate the settings themselves.
-  // If they choose yes, then automatically migrate the settings
-  // Lastly user can be directed to our README for more information about this
-  const choice = await vscode.window.showInformationMessage(
-    'Terraform Extension settings have moved in the latest update',
-    {
-      detail: messageText,
-      modal: false,
-    },
-    { title: 'More Info' },
-    { title: 'Migrate' },
-    { title: 'Open Settings' },
-    { title: 'Suppress' },
-  );
-  if (choice === undefined) {
-    return;
-  }
-
-  switch (choice.title) {
-    case 'Suppress':
-      ctx.globalState.update('terraform.disableSettingsMigration', true);
-      return;
-    case 'Open Settings':
-      await vscode.commands.executeCommand('workbench.action.openSettings', '@ext:hashicorp.terraform');
-      return;
-    case 'More Info':
-      await vscode.commands.executeCommand(
-        'vscode.open',
-        vscode.Uri.parse('https://github.com/hashicorp/vscode-terraform/blob/v2.24.0/docs/settings-migration.md'),
-      );
-      await migrateLegacySettings(ctx);
-      return;
-    case 'Migrate':
-    // migrate below
-  }
-
-  await migrate('terraform', 'languageServer.external', 'languageServer.enable');
-  await migrate('terraform', 'languageServer.pathToBinary', 'languageServer.path');
-
-  // We need to move args and ignoreSingleFileWarning out of the JSON object format
-  await migrate('terraform', 'languageServer.args', 'languageServer.args');
-  await migrate('terraform', 'languageServer.ignoreSingleFileWarning', 'languageServer.ignoreSingleFileWarning');
-  await deleteSetting('terraform', 'languageServer');
-
-  // This simultaneously moves terraform-ls to terraform as well as migrate setting names
-  await migrate('terraform-ls', 'rootModules', 'languageServer.rootModules');
-  await migrate('terraform-ls', 'excludeRootModules', 'languageServer.indexing.ignorePaths');
-  await migrate('terraform-ls', 'ignoreDirectoryNames', 'languageServer.indexing.ignoreDirectoryNames');
-  await migrate('terraform-ls', 'terraformExecPath', 'languageServer.terraform.path');
-  await migrate('terraform-ls', 'terraformExecTimeout', 'languageServer.terraform.timeout');
-  await migrate('terraform-ls', 'terraformLogFilePath', 'languageServer.terraform.logFilePath');
-
-  // We need to move prefillRequiredFields and validateOnSave out of the JSON object format as well as
-  // move terraform-ls to terraform
-  await migrate('terraform-ls', 'experimentalFeatures.validateOnSave', 'experimentalFeatures.validateOnSave');
-  await migrate(
-    'terraform-ls',
-    'experimentalFeatures.prefillRequiredFields',
-    'experimentalFeatures.prefillRequiredFields',
-  );
-  await deleteSetting('terraform-ls', 'experimentalFeatures');
-  await vscode.commands.executeCommand('workbench.action.reloadWindow');
-}
-
-function previewExtensionPresent(currentExtensionID: string) {
-  const stable = vscode.extensions.getExtension('hashicorp.terraform');
-  const preview = vscode.extensions.getExtension('hashicorp.terraform-preview');
-
-  const msg = 'Please ensure only one is enabled or installed and reload this window';
-
-  if (currentExtensionID === 'hashicorp.terraform-preview') {
-    if (stable !== undefined) {
-      vscode.window.showErrorMessage('Terraform Preview cannot be used while Terraform Stable is also enabled.' + msg);
-      return true;
-    }
-  } else if (currentExtensionID === 'hashicorp.terraform') {
-    if (preview !== undefined) {
-      vscode.window.showErrorMessage('Terraform Stable cannot be used while Terraform Preview is also enabled.' + msg);
-      return true;
-    }
-  }
-
-  return false;
 }
